requirements.txt
# Ensure all required packages for conversational memory and LLM API
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.4.2
jinja2==3.1.2

# LangChain and related packages
langchain>=0.0.310
langchain-community>=0.0.13
langchain-core>=0.1.9

# File processing
pymupdf==1.23.4
docx2txt==0.8
pytesseract==0.3.10
Pillow==10.0.1
unstructured==0.10.30
pdf2image==1.16.3

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Embedding and vector store
sentence-transformers==2.2.2
# faiss-cpu==1.7.4  # FAISS no longer required, can be removed if not used elsewhere
chromadb==0.4.18

# LLM integration
huggingface-hub==0.17.3
transformers==4.35.0

# Optional: For type checking and dev
# mypy
# pytest
